requirements/dev.txt
-r base.txt
django-debug-toolbar==1.9.1
django-querycount==0.7.0
